api/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
rembg
pillow
python-multipart
onnxruntime
